+++ src/icclim/models/index_group.py
@@ -52,5 +52,8 @@ class IndexGroupRegistry(Registry[IndexGroup]):
     RAIN = IndexGroup("rain")
     SNOW = IndexGroup("snow")
     WIND = IndexGroup("wind")
+    PRESSURE = IndexGroup("pressure")
+    SUNSHINE = IndexGroup("sunshine")
+    HUMIDITY = IndexGroup("humidity")
     # no climate index should be bounded to "all"
     WILD_CARD_GROUP = IndexGroup("all")